Search client by callsign and type
Some callers of this method rely on SearchClient to return only pointers
to the passed client type or null. In case there were two callsigns with
different types (e.g. pilot logged in as observer which defaults
to an ATC login), SearchClient could return a pointer to a Controller
object which is then casted to an Airplane object pointer.